Consolidate the downloading logic in a new pipes.fetch module #100
