+++ scanpipe/management/commands/add-input.py
@@ -43,56 +43,3 @@ class Command(AddInputCommandMixin, ProjectCommand):
 
         if input_urls:
             self.handle_input_urls(input_urls)
-
-    # def handle_input_files(self, inputs_files):
-    #     """
-    #     Copy provided `inputs_files` to the project `input` directory.
-    #     """
-    #     copied = []
-    #     for file_location in inputs_files:
-    #         self.project.copy_input_from(file_location)
-    #         filename = Path(file_location).name
-    #         copied.append(filename)
-    #         self.project.add_input_source(filename, source="uploaded")
-    #
-    #     msg = "File(s) copied to the project inputs directory:"
-    #     self.stdout.write(self.style.SUCCESS(msg))
-    #     msg = "\n".join(["- " + filename for filename in copied])
-    #     self.stdout.write(msg)
-    #
-    # @staticmethod
-    # def validate_input_files(inputs_files):
-    #     """
-    #     Raise an error if one of the provided `inputs_files` is not an existing file.
-    #     """
-    #     for file_location in inputs_files:
-    #         file_path = Path(file_location)
-    #         if not file_path.is_file():
-    #             raise CommandError(f"{file_location} not found or not a file")
-    #
-    # def handle_input_urls(self, input_urls):
-    #     """
-    #     Fetch provided `input_urls` and store to the project `input` directory.
-    #     """
-    #     downloads = []
-    #     errors = []
-    #
-    #     for url in input_urls:
-    #         try:
-    #             downloaded = download(url)
-    #             self.project.move_input_from(downloaded.file_path)
-    #             self.project.add_input_source(downloaded.filename, downloaded.uri)
-    #             downloads.append(downloaded)
-    #         except Exception:
-    #             errors.append(url)
-    #
-    #     if downloads:
-    #         msg = "File(s) downloaded to the project inputs directory:"
-    #         self.stdout.write(self.style.SUCCESS(msg))
-    #         msg = "\n".join(["- " + downloaded.filename for downloaded in downloads])
-    #         self.stdout.write(msg)
-    #
-    #     if errors:
-    #         self.stdout.write(self.style.ERROR("Could not fetch URL(s):"))
-    #         msg = "\n".join(["- " + url for url in errors])
-    #         self.stdout.write(self.style.ERROR(msg))
