+++ scanpipe/templates/scanpipe/project_form.html
@@ -49,10 +49,10 @@
               </div>
 
               <div class="field mt-5">
-                <label class="label has-text-weight-semibold" for="{{ form.download_urls.id_for_label }}">{{ form.download_urls.label }}</label>
+                <label class="label has-text-weight-semibold" for="{{ form.input_urls.id_for_label }}">{{ form.input_urls.label }}</label>
                 <div class="control">
-                  {{ form.download_urls }}
-                  <p class="help">{{ form.download_urls.help_text }}</p>
+                  {{ form.input_urls }}
+                  <p class="help">{{ form.input_urls.help_text }}</p>
                 </div>
               </div>
             </div>
